Some work

+++ llvm/aaa/simple_main/simple_main.c
@@ -1,3 +1,3 @@
-int main_lol() {
+unsigned short main_lol() {
   return 0x69;
 }
